+++ lang/en/format_grid.php
@@ -146,6 +146,10 @@ $string['sectionbadgeingridbox_help'] = 'Show the section badge in the grid box'
 $string['defaultsectionbadgeingridbox'] = 'Section badge in grid box';
 $string['defaultsectionbadgeingridbox_desc'] = 'Show the section badge in the grid box.';
 
+// Indentation.
+$string['indentation'] = 'Allow indentation on the course page';
+$string['indentation_help'] = 'Allow teachers and other users with the manage activities capability to indent items on the course page.';
+
 // Completion.
 $string['showcompletion'] = 'Show completion';
 $string['showcompletion_help'] = 'Show the completion percentage of the sections';